Conflicts:
	src/server/game/Achievements/AchievementMgr.cpp
	src/server/game/Spells/SpellInfo.cpp
	src/server/scripts/Commands/cs_misc.cpp

+++ src/server/game/Spells/SpellInfo.cpp
@@ -367,7 +367,7 @@ bool SpellEffectInfo::IsAura() const
 
 bool SpellEffectInfo::IsAura(AuraType aura) const
 {
-    return IsAura() && AuraType(ApplyAuraName) == aura;
+    return IsAura() && AuraType(ApplyAuraName) == uint32(aura);
 }
 
 bool SpellEffectInfo::IsTargetingArea() const
